move _wait_invocation to GalaxyTestBase
